refactor: move relevant categories components for categories folder
