+++ src/features/categories/relevant-categories/relevantCategories.css
@@ -1,0 +1,3 @@
+.frappe-chart {
+  min-height: 300px;
+}
